requirement.txt
pyradiomics
SimpleITK
numpy
scipy
pandas
scikit-learn
xgboost